Move java-time project into core
