+++ build.sbt
@@ -76,7 +76,7 @@ val commonSettings = Seq(
 )
 
 lazy val root = (project in file("."))
-  .aggregate(formats, scanamo, testkit, alpakka, refined, catsEffect, javaTime, joda, zio)
+  .aggregate(formats, scanamo, testkit, alpakka, refined, catsEffect, joda, zio)
   .settings(
     commonSettings,
     publishingSettings,
@@ -219,21 +219,6 @@ lazy val alpakka = (project in file("alpakka"))
     scalacOptions in (Compile, doc) += "-no-link-warnings"
   )
   .dependsOn(formats, scanamo, testkit % "test->test")
-
-lazy val javaTime = (project in file("java-time"))
-  .settings(
-    commonSettings,
-    publishingSettings,
-    name := "scanamo-time"
-  )
-  .settings(
-    libraryDependencies ++= List(
-      "org.scalatest"  %% "scalatest"                   % "3.0.8"  % Test,
-      "org.scalacheck" %% "scalacheck"                  % "1.13.5" % Test,
-      "com.47deg"      %% "scalacheck-toolbox-datetime" % "0.2.5"  % Test
-    )
-  )
-  .dependsOn(formats)
 
 lazy val joda = (project in file("joda"))
   .settings(
